[level 0] Title: 세균 증식, Time: 0.10 ms, Memory: 87.5 MB -BaekjoonHub

+++ 프로그래머스/0/120910. 세균 증식/세균 증식.java
@@ -1,6 +1,6 @@
 class Solution {
-    public double solution(int n, int t) {
-        double answer = n * Math.pow(2, t);
+    public int solution(int n, int t) {
+        int answer = n * (int) Math.pow(2, t);
         return answer;
     }
 }
